[AEGIS-4898] produce csvs for tracking last run time
